+++ zinc/error.h
@@ -41,7 +41,7 @@ ZINC_API bool error_list_set(struct error_list *el, struct location* loc, const 
 ZINC_API void error_list_print(struct error_list* el);
 ZINC_API void location_init(struct location* loc);
 ZINC_API void location_create(struct location** loc);
-ZINC_API const char* plural(int number);
+ZINC_API const char* plural(size_t number);
 
 #ifdef __cplusplus
 }
